backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
dspy-ai==2.3.4
python-dotenv==1.0.0
python-multipart==0.0.9
librosa==0.10.1
soundfile==0.12.1
transformers==4.37.2
torch==2.2.0
Pillow==10.2.0
PyMuPDF==1.23.26
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.0